fixes #41 Test coverage for entity model  -  Removed original class tests
